+++ Recipes.BLL/Services/RecipeService.cs
@@ -1,12 +1,10 @@
 using AutoMapper;
-using Microsoft.EntityFrameworkCore.Query.Internal;
-using Recipes.BLL.Interfaces;
-using Recipes.DAL.Interfaces;
+using Recipes.BLL.Helpers;
+using Recipes.BLL.Services.Interfaces;
+using Recipes.DAL.Infrastructure.Interfaces;
 using Recipes.Data.DataTransferObjects;
-using Recipes.Data.Enums;
-using Recipes.Data.Interfaces;
 using Recipes.Data.Models;
-using Recipes.Data.Responses;
+using Recipes.Data.Responses.Interfaces;
 
 namespace Recipes.BLL.Services;
 
@@ -14,50 +12,51 @@ public class RecipeService : IRecipeService
 {
     private readonly IUnitOfWork _unitOfWork;
     private readonly IMapper _mapper;
+    private readonly ResponseCreator _responseCreator;
 
     public RecipeService(IUnitOfWork unitOfWork, IMapper mapper)
     {
         _unitOfWork = unitOfWork;
         _mapper = mapper;
+        _responseCreator = new ResponseCreator();
     }
 
     public async Task<IBaseResponse<RecipeDto>> GetById(Guid id)
     {
         try
         {
-            var recipe = await _unitOfWork.RecipeRepository.GetByIdAsync(id);
-            if (recipe == null)
-            {
-                return BaseResponse<RecipeDto>.CreateBaseResponse<RecipeDto>($"The recipe with id {id} wasn't found", StatusCode.NotFound);
-            }
+            if (id == Guid.Empty)
+                return _responseCreator.CreateBaseBadRequest<RecipeDto>("Id is empty.");
+            
+            var recipeDto = _mapper.Map<RecipeDto>(await _unitOfWork.RecipeRepository.GetByIdAsync(id));
+            
+            if (recipeDto == null)
+                return _responseCreator.CreateBaseNotFound<RecipeDto>($"Recipe with id {id} not found.");
 
-            return BaseResponse<RecipeDto>.CreateBaseResponse<RecipeDto>("Sucess!", StatusCode.Ok, _mapper.Map<RecipeDto>(recipe));
+            return _responseCreator.CreateBaseOk(recipeDto, 1);
         }
         catch (Exception ex)
         {
-            return BaseResponse<RecipeDto>.CreateBaseResponse<RecipeDto>(ex.Message, StatusCode.InternalServerError);
-
+            return _responseCreator.CreateBaseServerError<RecipeDto>(ex.Message);
         }
     }
 
-    public async Task<IBaseResponse<IEnumerable<RecipeDto>>> Get()
+    public async Task<IBaseResponse<List<RecipeDto>>> Get()
     {
         try
         {
             var models = await _unitOfWork.RecipeRepository.GetAsync();
 
             if (models.Count is 0)
-            {
-                return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>("0 objects found", StatusCode.NotFound);
-            }
+                return _responseCreator.CreateBaseNotFound<List<RecipeDto>>("No recipes found.");
 
             var dtoList = models.Select(model => _mapper.Map<RecipeDto>(model)).ToList();
 
-            return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>("Success!", StatusCode.Ok, dtoList, dtoList.Count);
+            return _responseCreator.CreateBaseOk(dtoList, dtoList.Count);
         }
         catch (Exception e)
         {
-            return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>(e.Message, StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<List<RecipeDto>>(e.Message);
         }
     }
 
@@ -66,32 +65,29 @@ public class RecipeService : IRecipeService
         try
         {
             if (modelDto is null) 
-                return BaseResponse<RecipeDto>.CreateBaseResponse<string>("Objet can`t be empty...", StatusCode.BadRequest);
+                return _responseCreator.CreateBaseBadRequest<string>("Inserted recipe is empty.");
             
             modelDto.Id = Guid.NewGuid();
 
             await _unitOfWork.RecipeRepository.InsertAsync(_mapper.Map<Recipe>(modelDto));
             await _unitOfWork.SaveChangesAsync();
 
-            return BaseResponse<RecipeDto>.CreateBaseResponse<string>("Object inserted!", StatusCode.Ok, resultsCount: 1);
-
+            return _responseCreator.CreateBaseOk($"Recipe added.", 1);
         }
         catch (Exception e)
         {
-            return BaseResponse<RecipeDto>.CreateBaseResponse<string>(e.Message, StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<string>(e.Message);
         }
     }
-
-
-    public async Task<IBaseResponse<string>> InsertWithIngredients(RecipeDtoWithIngredientsAndSteps? recipereqest)
+    
+    public async Task<IBaseResponse<string>> InsertWithIngredients(RecipeDtoWithIngredientsAndSteps? recipeRequest)
     {
         try
         {
-            if (recipereqest is null)
-                return BaseResponse<RecipeDto>.CreateBaseResponse<string>("Object can't be empty...", StatusCode.BadRequest);
-
-
-            var recipe = _mapper.Map<Recipe>(recipereqest);
+            if (recipeRequest is null)
+                return _responseCreator.CreateBaseBadRequest<string>("Inserted recipe is empty.");
+            
+            var recipe = _mapper.Map<Recipe>(recipeRequest);
 
             recipe.Id = Guid.NewGuid();
 
@@ -107,24 +103,15 @@ public class RecipeService : IRecipeService
                 dbIng.WeightUnit.Type == dtoIng.WeightUnit.Type &&
                 dbIng.Quantity == dtoIng.Quantity));
 
-            var steps = _mapper.Map<ICollection<CookingStepDtoNoId>, ICollection<CookingStep>>(recipereqest.Steps);
-
+            var steps = _mapper.Map<ICollection<CookingStepDtoNoId>, ICollection<CookingStep>>(recipeRequest.Steps);
             
             foreach(var step in steps)
             {
-
                 step.Id = Guid.NewGuid();
-
                 step.RecipeId = recipe.Id;
                 await _unitOfWork.CookingStepRepository.InsertAsync(step);
-
-
             }
-
-
-
-
-
+            
             // Додати нові інгредієнти в базу даних
             foreach (var newIngredient in newIngredients)
             {
@@ -134,17 +121,15 @@ public class RecipeService : IRecipeService
 
                 if (unit == null)
                 {
-               
                     await _unitOfWork.WeightUnitRepository.InsertAsync(_mapper.Map<WeightUnit>(newIngredient.WeightUnit));
                     //await _unitOfWork.SaveChangesAsync();
                 }
                 else
                     newIngredient.WeightUnit = unit;
-              await  _unitOfWork.IngredientRepository.InsertAsync(newIngredient);
+                await  _unitOfWork.IngredientRepository.InsertAsync(newIngredient);
             }
 
             // Перезаписати Id інгредієнтів, які вже існують в базі даних
-      
             foreach (var existingIngredient in existingIngredients)
             {
                 var recipesIngredient = recipe.Ingredients.FirstOrDefault(p =>
@@ -155,8 +140,6 @@ public class RecipeService : IRecipeService
                 if (recipesIngredient != null)
                 {
                     // Оновлюємо властивості об'єкта, який вже відстежується контекстом
-                  
-
                      var unit = _unitOfWork.WeightUnitRepository.GetAsync().Result.FirstOrDefault(p => p.Type == recipesIngredient.WeightUnit.Type);
                     
                     if (unit == null)
@@ -167,63 +150,62 @@ public class RecipeService : IRecipeService
                     {
                         recipesIngredient.WeightUnit = unit;
                     }
-
                     
                     recipe.Ingredients.Remove(recipesIngredient);
                     recipe.Ingredients.Add(existingIngredient);
                 }
             }
-
-
+            
             // Вставити рецепт в базу даних
             await _unitOfWork.RecipeRepository.InsertAsync(recipe);
             await _unitOfWork.SaveChangesAsync();
 
-            return BaseResponse<RecipeDto>.CreateBaseResponse<string>("Object inserted!", StatusCode.Ok, resultsCount: 1);
+            return _responseCreator.CreateBaseOk($"Recipe added.", 1);
         }
         catch (Exception e)
         {
-            return BaseResponse<RecipeDto>.CreateBaseResponse<string>(e.Message, StatusCode.InternalServerError, recipereqest.ToString());
+            return _responseCreator.CreateBaseServerError<string>(e.Message);
         }
     }
     public async Task<IBaseResponse<string>> DeleteById(Guid id)
     {
         try
         {
+            if (id == Guid.Empty)
+                return _responseCreator.CreateBaseBadRequest<string>("Id is empty.");
+            
             await _unitOfWork.RecipeRepository.DeleteAsync(id);
             await _unitOfWork.SaveChangesAsync();
 
-            return BaseResponse<RecipeDto>.CreateBaseResponse<string>("Object deleted!", StatusCode.Ok, resultsCount: 1);
+            return _responseCreator.CreateBaseOk("Recipe deleted.", 1);
         }
         catch (Exception e)
         {
-            return BaseResponse<RecipeDto>.CreateBaseResponse<string>($"{e.Message} or object not found", StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<string>(e.Message);
         }
     }
-
-
-
-    public async Task<IBaseResponse<IEnumerable<RecipeDto>>> GetByName(string name)
+    
+    public async Task<IBaseResponse<List<RecipeDto>>> GetByName(string name)
     {
         try
         {
+            if (string.IsNullOrEmpty(name))
+                return _responseCreator.CreateBaseBadRequest<List<RecipeDto>>("Name is empty.");
+            
             var models = await _unitOfWork.RecipeRepository.GetAsync();
 
             if (models.Count == 0)
-            {
-                return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>("0 objects found", StatusCode.NotFound);
-            }
+                return _responseCreator.CreateBaseNotFound<List<RecipeDto>>("No recipes found.");
 
-            var dtoList = models.Select(model => _mapper.Map<RecipeDto>(model)).ToList().Where(x => x.Title.Contains(name)); ;
+            var dtoList = models.Select(model => _mapper.Map<RecipeDto>(model)).ToList().Where(x => x.Title.Contains(name)).ToList(); ;
 
-            return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>("Success!", StatusCode.Ok, dtoList, dtoList.ToList().Count);
+            return _responseCreator.CreateBaseOk(dtoList, dtoList.Count);
         }
         catch (Exception e)
         {
-            return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>(e.Message, StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<List<RecipeDto>>(e.Message);
         }
     }
-
 
     /*    public async Task<IBaseResponse<IEnumerable<RecipeDto>>> GetByIngredients(IEnumerable<string> ingredients)
         {
@@ -326,49 +308,48 @@ public class RecipeService : IRecipeService
                 return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>(e.Message, StatusCode.InternalServerError);
             }
         }*/
-    public async Task<IBaseResponse<IEnumerable<RecipeDto>>> GetByIngredients(IEnumerable<string> ingredients)
+    
+    public async Task<IBaseResponse<List<RecipeDto>>> GetByIngredients(IEnumerable<string> ingredients)
     {
         try
         {
+            var enumerable = ingredients.ToList();
+            if (!enumerable.ToList().Any())
+                return _responseCreator.CreateBaseBadRequest<List<RecipeDto>>("No ingredients specified.");
+            
             var models = await _unitOfWork.RecipeRepository.GetAsync();
 
             if (models.Count == 0)
-            {
-                return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>("0 objects found", StatusCode.NotFound);
-            }
+                return _responseCreator.CreateBaseNotFound<List<RecipeDto>>("No recipes found.");
 
             // Отримання інгредієнтів з бази даних
             var allIngredients = await _unitOfWork.IngredientRepository.GetAsync();
 
             // Фільтрація інгредієнтів зі списку ingredients
-            var filteredIngredients = allIngredients.Where(ing => ingredients.Contains(ing.Name)).ToList();
+            var filteredIngredients = allIngredients.Where(ing => enumerable.Contains(ing.Name)).ToList();
 
             // Отримання рецептів, які містять хоча б один інгредієнт зі списку ingredients, але не містять рецептів без інгредієнтів зі списку
             var matchingRecipes = new List<Recipe>();
 
             foreach (var model in models)
             {
-                if (model.Ingredients.Any(ing => ingredients.Contains(ing.Name)) && !model.Ingredients.All(ing => !ingredients.Contains(ing.Name)))
+                if (model.Ingredients.Any(ing => enumerable.Contains(ing.Name)) && !model.Ingredients.All(ing => !enumerable.Contains(ing.Name)))
                 {
                     matchingRecipes.Add(model);
                 }
             }
 
             if (matchingRecipes.Count == 0)
-            {
-                return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>("No recipes found with any of the specified ingredients", StatusCode.NotFound);
-            }
-
+                return _responseCreator.CreateBaseNotFound<List<RecipeDto>>("No recipes found with any of the specified ingredients.");
+            
             // Мапування знайдених рецептів на DTO
             var dtoList = matchingRecipes.Select(model => _mapper.Map<RecipeDto>(model)).ToList();
 
-            return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>("Success!", StatusCode.Ok, dtoList, matchingRecipes.Count);
+            return _responseCreator.CreateBaseOk(dtoList, dtoList.Count);
         }
         catch (Exception e)
         {
-            return BaseResponse<RecipeDto>.CreateBaseResponse<IEnumerable<RecipeDto>>(e.Message, StatusCode.InternalServerError);
+            return _responseCreator.CreateBaseServerError<List<RecipeDto>>(e.Message);
         }
     }
-
-
 }
